request types in passport-interface
